Move tasks to notify
All the 'tasks' were about notifying users.
